+++ mapchete/commons/hillshade.py
@@ -86,29 +86,40 @@ def calculate_slope_aspect(elevation, xres, yres, z=1.0, scale=1.0):
     return slope, aspect
 
 
-def hillshade(elevation, tile, azimuth=315.0, altitude=45.0, z=1.0, scale=1.0, ):
+def hillshade(
+    elevation,
+    tile,
+    azimuth=315.0,
+    altitude=45.0,
+    z=1.0,
+    scale=1.0,
+):
     """
     Return hillshaded numpy array.
 
     Parameters
     ----------
     elevation : array
-        input elevation data
+        Input elevation data.
     tile : Tile
-        tile covering the array
+        Tile covering the array.
+    azimuth : float
+        Light source direction in degrees. (default: 315, top left)
+    altitude : float
+        Light source altitude angle in degrees. (default: 45)
     z : float
-        vertical exaggeration factor
+        Vertical DEM exaggeration factor. (default: 1)
     scale : float
-        scale factor of pixel size units versus height units (insert 112000
-        when having elevation values in meters in a geodetic projection)
+        Scale factor of pixel size units versus height units (insert 112000
+        when having elevation values in meters in a geodetic projection).
     """
     elevation = elevation[0] if elevation.ndim == 3 else elevation
     azimuth = float(azimuth)
     altitude = float(altitude)
     z = float(z)
     scale = float(scale)
-    xres = tile.tile.pixel_x_size
-    yres = -tile.tile.pixel_y_size
+    xres = tile.pixel_x_size
+    yres = -tile.pixel_y_size
     slope, aspect = calculate_slope_aspect(
         elevation,
         xres,
@@ -117,13 +128,16 @@ def hillshade(elevation, tile, azimuth=315.0, altitude=45.0, z=1.0, scale=1.0, )
         scale=scale
     )
     deg2rad = math.pi / 180.0
+    # shaded has values between -1.0 and +1.0
     shaded = np.sin(altitude * deg2rad) * np.sin(slope) \
         + np.cos(altitude * deg2rad) * np.cos(slope) \
         * np.cos((azimuth - 90.0) * deg2rad - aspect)
-    # shaded now has values between -1.0 and +1.0
-    # stretch to 0 - 255 and invert
-    shaded = (((shaded + 1.0) / 2) * -255.0).astype("uint8")
-    # add one pixel padding using the edge values
+    # stretch to 0 - 255 and add one pixel padding using the edge values
     return ma.masked_array(
-        data=np.pad(shaded, 1, mode='edge'), mask=elevation.mask
+        data=np.pad(
+            np.clip(shaded * 255.0, 1, 255).astype("uint8"),
+            1,
+            mode='edge'
+        ),
+        mask=elevation.mask
     )