+++ mapchete/formats/default/gtiff.py
@@ -52,7 +52,7 @@ from mapchete.io.raster import (
     extract_from_array, read_raster_window
 )
 from mapchete.tile import BufferedTile
-from mapchete._validate import deprecated_kwargs
+from mapchete.validate import deprecated_kwargs
 
 
 logger = logging.getLogger(__name__)
@@ -286,9 +286,12 @@ class GTiffTileDirectoryOutputReader(
         profile = self.profile(process_tile)
         return ma.masked_array(
             data=np.full(
-                (profile["count"], ) + process_tile.shape, profile["nodata"],
-                dtype=profile["dtype"]),
-            mask=True
+                (profile["count"], ) + process_tile.shape,
+                profile["nodata"],
+                dtype=profile["dtype"]
+            ),
+            mask=True,
+            fill_value=profile["nodata"]
         )
 
     def profile(self, tile=None):