+++ test/test_processes.py
@@ -5,7 +5,7 @@ import numpy.ma as ma
 
 import mapchete
 from mapchete.processes.examples import example_process
-from mapchete.processes import convert
+from mapchete.processes import contours, convert, hillshade
 
 
 def test_example_process(cleantopo_tl):
@@ -85,3 +85,109 @@ def test_convert(cleantopo_tl, cleantopo_tl_tif, landpoly):
             input=mp.config.get_inputs_for_tile(tile),
         )
         assert convert.execute(user_process) == "empty"
+
+
+def test_contours(cleantopo_tl, cleantopo_tl_tif, landpoly):
+    with mapchete.open(
+        dict(cleantopo_tl.dict, input=dict(dem=cleantopo_tl_tif))
+    ) as mp:
+        zoom = max(mp.config.zoom_levels)
+        # execute without clip
+        tile = next(mp.get_process_tiles(zoom))
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        output = contours.execute(user_process)
+        assert isinstance(output, list)
+        assert output
+        # execute on empty tile
+        tile = mp.config.process_pyramid.tile(
+            zoom,
+            mp.config.process_pyramid.matrix_height(zoom) - 1,
+            mp.config.process_pyramid.matrix_width(zoom) - 1
+        )
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert contours.execute(user_process) == "empty"
+
+    with mapchete.open(
+        dict(cleantopo_tl.dict, input=dict(dem=cleantopo_tl_tif, clip=landpoly))
+    ) as mp:
+        zoom = max(mp.config.zoom_levels)
+        tile = next(mp.get_process_tiles(zoom))
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        output = contours.execute(user_process)
+        assert isinstance(output, list)
+        assert output
+        # execute on empty tile
+        tile = mp.config.process_pyramid.tile(
+            zoom,
+            mp.config.process_pyramid.matrix_height(zoom) - 1,
+            mp.config.process_pyramid.matrix_width(zoom) - 1
+        )
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert contours.execute(user_process) == "empty"
+
+
+def test_hillshade(cleantopo_tl, cleantopo_tl_tif, landpoly):
+    with mapchete.open(
+        dict(cleantopo_tl.dict, input=dict(dem=cleantopo_tl_tif))
+    ) as mp:
+        zoom = max(mp.config.zoom_levels)
+        # execute without clip
+        tile = next(mp.get_process_tiles(zoom))
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert isinstance(hillshade.execute(user_process), np.ndarray)
+        # execute on empty tile
+        tile = mp.config.process_pyramid.tile(
+            zoom,
+            mp.config.process_pyramid.matrix_height(zoom) - 1,
+            mp.config.process_pyramid.matrix_width(zoom) - 1
+        )
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert hillshade.execute(user_process) == "empty"
+
+    with mapchete.open(
+        dict(cleantopo_tl.dict, input=dict(dem=cleantopo_tl_tif, clip=landpoly))
+    ) as mp:
+        zoom = max(mp.config.zoom_levels)
+        tile = next(mp.get_process_tiles(zoom))
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert isinstance(hillshade.execute(user_process), np.ndarray)
+        # execute on empty tile
+        tile = mp.config.process_pyramid.tile(
+            zoom,
+            mp.config.process_pyramid.matrix_height(zoom) - 1,
+            mp.config.process_pyramid.matrix_width(zoom) - 1
+        )
+        user_process = mapchete.MapcheteProcess(
+            tile=tile,
+            params=mp.config.params_at_zoom(tile.zoom),
+            input=mp.config.get_inputs_for_tile(tile),
+        )
+        assert hillshade.execute(user_process) == "empty"